+++ spec/bali/components/reveal_spec.rb
@@ -41,7 +41,7 @@ RSpec.describe Bali::Reveal::Component, type: :component do
 
   it 'does not render border at bottom' do
     render_inline(component) do |c|
-      c.trigger(title: 'Click here', border: false)
+      c.trigger(title: 'Click here', show_border: false)
     end
 
     render_inline(component)